# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是详细介绍待办事项的 API 文档",
        "title": "待办事项 API 文档",
        "contact": {
            "name": "huang",
            "email": "nanguatou10@gmail"
        },
        "version": "1.0"
    },
    "paths": {
        "/ping": {
            "get": {
                "description": "返回 pong",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "示例"
                ],
                "summary": "测试接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/todo": {
            "get": {
                "description": "返回给前端所有的 Todo 项目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "查询所有待办事项",
                "responses": {
                    "200": {
                        "description": "返回所有待办事项",
                        "schema": {
                            "$ref": "#/definitions/models.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "接收前端传来的 JSON，创建一个 Todo 项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "创建待办事项",
                "parameters": [
                    {
                        "description": "待办事项内容",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功返回的结构体",
                        "schema": {
                            "$ref": "#/definitions/models.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/todo/{id}": {
            "put": {
                "description": "根据 ID 更新待办事项的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "修改待办事项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "待办事项id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "待办事项内容",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功返回的结构体",
                        "schema": {
                            "$ref": "#/definitions/models.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据 ID 删除待办事项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todo"
                ],
                "summary": "删除待办事项",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "待办事项id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功返回的结构体",
                        "schema": {
                            "$ref": "#/definitions/models.TodoResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Todo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.TodoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Todo"
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}